move load font into own file
