+++ packages/cli/src/process-image.ts
@@ -2,18 +2,7 @@
 
 import Jimp = require('jimp');
 import { log, greenCheck } from './log';
-
-export async function loadFont(
-  font?: string,
-  verbose?: boolean
-): Promise<Jimp.Font> {
-  if (font) {
-    log(` 🔤  Loading font: ${font} ...`, verbose);
-    return await Jimp.loadFont(Jimp[font] || font);
-  }
-
-  return;
-}
+import { loadFont } from './load-font';
 
 function runActions(
   image: Jimp,
